feat: Support providing both a context and an error object
This PR solves the problem where you need to manually create a context object from the error context
and the error object  before passing it to `logger.error`.

You can now provide both the error object and the context in which it happened to `logger.error`
and it will merge both objects automatically.

Example:
```js
const a = 1;
const b = 2;

try {
 doSomething(a, b);
}
catch (error)
  logger.error('Failed to process the data', {a, b}, error);
}
```